backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "URL Ingestion & Vector Storage System - Ingests website URLs from sitemap, generates embeddings with Cohere, and stores in Qdrant cloud for RAG applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0"
]
